Renamed json parser -> provider
